+++ docs/source/_static/js/custom.js
@@ -1,5 +1,12 @@
+
+
 window.addEventListener('load', function() {
    console.log("Custom JS file loaded");
+   let operationContainers = document.querySelectorAll(".container.operations");
+   
+   for (let opContainer of operationContainers) {
+      opContainer.innerHTML = `<b>Supported operations:</b><br/> ${opContainer.innerHTML}`
+   }
    // const classes = document.getElementsByClassName("class");
    let tables = document.getElementsByClassName('py-attribute-table');