+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # 📜 Weatherly changelog
 In this file you can view the changelog, including updates and changes that were made to the package.
+ 
+### Version 0.4.0
+This version adds support for History API with some new, cool features
+* New exception: `weatherly.InvalidDate`, raised from `weatherly.get_historical_data` when the date is invalid
+* `weatherly.get_historical_data` : you can now get historical forecast data as `ForecastData` object
+* `weatherly.ForecastData.iter_hours` : an iterator that lets you easily iterate over every hour data from the class.
+
+Improvements to the documentation:
+* Added `.. container:: operations` handling with custom CSS and JS. This directive is used for desribing operations for a class, function etc.
 
 ### Version 0.3.0
 This version adds support for Forecast API with a lot of new features and classes.